pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krauler"
dynamic = ["version"]
description = "🚀🤖 krauler: Clean, AI-ready web crawler for invention pipelines"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "StadiumPhenom", email = "you@example.com"}
]
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.11",
    "aiosqlite~=0.20",
    "anyio>=4.0.0",
    "lxml~=5.3",
    "litellm>=1.53.1",
    "numpy>=1.26.0,<3",
    "pillow>=10.4",
    "playwright>=1.49.0",
    "patchright>=1.49.0",
    "python-dotenv~=1.0",
    "requests~=2.26",
    "beautifulsoup4~=4.12",
    "tf-playwright-stealth>=1.1.0",
    "xxhash~=3.4",
    "rank-bm25~=0.2",
    "snowballstemmer~=2.2",
    "pydantic>=2.10",
    "pyOpenSSL>=24.3.0",
    "psutil>=6.1.1",
    "PyYAML>=6.0",
    "nltk>=3.9.1",
    "rich>=13.9.4",
    "httpx[http2]>=0.27.2",
    "fake-useragent>=2.0.3",
    "click>=8.1.7",
    "chardet>=5.2.0",
    "brotli>=1.1.0",
    "humanize>=4.10.0",
    "lark>=1.2.2",
    "alphashape>=1.3.1",
    "shapely>=2.0.0",
    "streamlit>=1.39.0",
    "pandas>=2.2.0",
    "altair>=5.3.0"
]

[project.optional-dependencies]
pdf = ["pypdf", "pdfminer.six"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
